tests/wpt/tests/visual-viewport/viewport-unscaled-scale.html

<!doctype html>
<html>
    <head>
        <title>Viewport: Unscaled scale</title>
        <meta charset="utf-8">
        <meta name="viewport" content="width=device-width">
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="viewport_support.js"></script>
    </head>
    <body>
    <h1>Viewport: Unscaled scale</h1>
    <h4>Test Description: This test checks that the default value for scale is 1.</h4>
    <div id="complete-notice">
        <p>window.visualViewport.scale is <span id="view-scale-log"></span>.</p>
    </div>
    <div id="log"></div>
    </body>
    <script>
        test(function() {
              assert_equals(window.visualViewport.scale, 1);
            }, "visualViewport.scale default value.");

        document.getElementById("view-scale-log").innerText = window.visualViewport.scale;
    </script>
</html>
