third_party/blink/web_tests/platform/android/external/wpt/visual-viewport/viewport-unscaled-scale-expected.txt

This is a testharness.js-based test.
[FAIL] visualViewport.scale default value.
  assert_equals: expected 1 but got 0.36734694242477417
Harness: the test ran to completion.
